Create func for common template context entries
